+++ .gitignore
@@ -10,6 +10,7 @@
 ###############
 bin/
 vendor/
+node_modules/
 
 ### Files ###
 #############
