+++ certificate_test.go
@@ -1,3 +1,5 @@
+// +build !js
+
 package webrtc
 
 import (
