+++ datachannel_js.go
@@ -127,13 +127,23 @@ func (d *DataChannel) Label() string {
 // Ordered represents if the DataChannel is ordered, and false if
 // out-of-order delivery is allowed.
 func (d *DataChannel) Ordered() bool {
-	return d.underlying.Get("ordered").Bool()
+	ordered := d.underlying.Get("ordered")
+	if ordered == js.Undefined() {
+		return true // default is true
+	}
+	return ordered.Bool()
 }
 
 // MaxPacketLifeTime represents the length of the time window (msec) during
 // which transmissions and retransmissions may occur in unreliable mode.
 func (d *DataChannel) MaxPacketLifeTime() *uint16 {
-	return valueToUint16Pointer(d.underlying.Get("maxPacketLifeTime"))
+	if d.underlying.Get("maxPacketLifeTime") != js.Undefined() {
+		return valueToUint16Pointer(d.underlying.Get("maxPacketLifeTime"))
+	} else {
+		// See https://bugs.chromium.org/p/chromium/issues/detail?id=696681
+		// Chrome calls this "maxRetransmitTime"
+		return valueToUint16Pointer(d.underlying.Get("maxRetransmitTime"))
+	}
 }
 
 // MaxRetransmits represents the maximum number of retransmissions that are
