+++ package.json
@@ -1,0 +1,8 @@
+{
+  "name": "webrtc",
+  "repository": "git@github.com:pions/webrtc.git",
+  "private": true,
+  "devDependencies": {
+    "wrtc": "^0.3.5"
+  }
+}
